+++ quelware-client/src/quelware_client/infra/diagnostics_agent_grpc/__init__.py
@@ -3,6 +3,7 @@ from grpclib.client import Channel
 from quelware_core.entities.resource import ResourceId
 
 from quelware_client.core.interfaces.diagnostics_agent import DiagnosticsAgent
+from quelware_client.infra._grpc_retry import call_with_retry
 
 
 class DiagnosticsAgentGrpc(DiagnosticsAgent):
@@ -12,7 +13,9 @@ class DiagnosticsAgentGrpc(DiagnosticsAgent):
 
     async def dump_port_state(self, port_id: ResourceId) -> str:
         req = pb_diag.DumpPortStateRequest(port_id=str(port_id))
-        resp = await self._service.dump_port_state(req)
+        resp = await call_with_retry(
+            lambda: self._service.dump_port_state(req), idempotent=True
+        )
         return resp.text